requirements.txt
supabase
chromadb
groq
python-dotenv
waitress
sentence-transformers
pyngrok 
flask
ngrok
pandas
PyPDF2
python-docx
openpyxl
pytest
opencv-python
scikit-image
pytesseract
numpy
pillow
requests
aiohttp
beautifulsoup4
python-magic
hypercorn 
asgiref
plotly
dash
dash-bootstrap-components
scikit-learn
matplotlib
seaborn
